+++ ruuvitag_sensor/common.py
@@ -1,0 +1,9 @@
+class RunFlag(object):
+    """
+    Wrapper for boolean run flag
+
+    Attributes:
+        running (bool): Defines if function should continue execution
+    """
+
+    running = True